+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 + Added Career entity and an endpoint
 + Added Test too
 
++ [[Backend-374]](https://www.notion.so/Talle-M-Ver-detalle-de-una-reuni-n-172ff13786bd405fade53cf161a723ee?pvs=4)
++ Removed description and mode from Meet model.
++ Adjusted /meets/:id response payload
+
 + [[Backend-354]](https://www.notion.so/Talle-S-Verificar-l-gica-de-interesarse-desinteresarse-en-una-meet-2b8bf228b0e84306abb8707ed5ce06dc)
 + Fixed logic to interest and uninterest in a meet.
 + New endpoint POST /meets/:id/interesteds to interest in a meet.